requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
langchain
langchain-community
openai==1.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
langchain_openai
